Fix Preserve Maven goal wiring and functional tests
