+++ native-maven-plugin/docs/e2e.md
@@ -104,8 +104,8 @@ valid layer.
 
 ### 3.9 Dependency preservation
 
-`PreserveDependencyFunctionalTest` verifies that an application compile goal resolves a configured
-dependency and its transitive artifacts, emits one path-based Preserve option, and retains a class
+`PreserveDependencyFunctionalTest` verifies that an application compile goal selects one specific
+transitive library non-transitively, emits one path-based Preserve option, and retains a class
 which an equivalent control image cannot load. This protects [§FS-native-builds.3](functional/native-image-builds.md#3-classpath-and-scopes),
 [§FS-config-model.8](functional/configuration-model.md#8-preserve-dependency-selection), and
 [§FS-goal-surface.1](functional/goal-surface.md#1-build-goals).
