+++ native-maven-plugin/src/test/groovy/org/graalvm/buildtools/maven/PreserveConfigurationTest.groovy
@@ -44,7 +44,7 @@ import org.graalvm.buildtools.maven.config.PreserveConfiguration
 import org.graalvm.buildtools.maven.config.PreserveDependencyConfiguration
 import spock.lang.Specification
 
-// Verifies Maven XML bean semantics and application-goal ownership for Preserve. §FS-config-model.8.
+// Verifies Maven XML bean semantics and compile-no-fork hierarchy ownership for Preserve. §FS-config-model.8.
 class PreserveConfigurationTest extends Specification {
     def "dependency selection is transitive by default and configurable"() {
         given:
@@ -65,17 +65,14 @@ class PreserveConfigurationTest extends Specification {
         PreserveConfiguration.declaredFields.findAll { !it.synthetic }*.name == ["dependencies"]
     }
 
-    def "parameter belongs only to application compile goals"() {
+    def "parameter belongs to the compile-no-fork hierarchy"() {
         expect:
-        NativeCompileNoForkGoalMojo.getDeclaredField("preserve")
-        NativeCompileNoForkGoalMojo.superclass == NativeCompileNoForkMojo
-        NativeCompileMojo.getDeclaredField("preserve")
-        DeprecatedNativeBuildMojo.getDeclaredField("preserve")
+        NativeCompileNoForkMojo.getDeclaredField("preserve")
         NativeCompileMojo.superclass == NativeCompileNoForkMojo
         DeprecatedNativeBuildMojo.superclass == NativeCompileNoForkMojo
+        WriteArgsFileMojo.superclass == NativeCompileNoForkMojo
         AbstractNativeImageMojo.declaredMethods*.name.contains("preserveConfiguration")
         !AbstractNativeImageMojo.declaredMethods*.name.contains("getPreserveConfiguration")
-        !WriteArgsFileMojo.declaredFields*.name.contains("preserve")
         !NativeTestMojo.declaredFields*.name.contains("preserve")
         !NativeIntegrationTestMojo.declaredFields*.name.contains("preserve")
         !LayerCreateMojo.declaredFields*.name.contains("preserve")
